sourceStringValidation & escapeRegex tested as one
