--- src/UI/api/utils/escapeRegex.ts
@@ -1,11 +1,0 @@
-export function escapeRegex(inputField: string) {
-  if (typeof inputField === 'boolean') {
-    return inputField;
-  } else {
-    let escapeinputField = inputField.replace(
-      /[-[\]/{}()*+?.,\\^$|#"]/g,
-      '\\$&'
-    );
-    return escapeinputField;
-  }
-}
